requirements.txt
pandas
streamlit
Pillow
openpyxl
numpy
